refocus on code proposal
